# openapi.yaml
openapi: 3.0.0
paths:
...
  /organizations:
    post:
      operationId: OrganizationController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationDto'
      responses:
        '201':
          description: Organization created successfully
        '403':
          description: Admin role required
        '409':
          description: Organization with this name or slug already exists
      security:
        - bearer: []
      summary: Create a new organization (admin only)
      tags:
        - Organizations
    get:
      operationId: OrganizationController_findAll
      parameters: []
      responses:
        '200':
          description: Array of all organizations
        '403':
          description: Admin role required
      security:
        - bearer: []
      summary: List all organizations (admin only)
      tags:
        - Organizations
  /organizations/{id}:
    get:
      operationId: OrganizationController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Organization ID
          schema:
            type: string
      responses:
        '200':
          description: Organization details
        '404':
          description: Organization not found
      security:
        - bearer: []
      summary: Get an organization by ID (must be the caller's own org, unless admin)
      tags:
        - Organizations
    patch:
      operationId: OrganizationController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Organization ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationDto'
      responses:
        '200':
          description: Updated organization
        '403':
          description: Only an admin can change the default organization
        '404':
          description: Organization not found
        '409':
          description: Name or slug conflict
      security:
        - bearer: []
      summary: Update an organization (must be the caller's own org, unless admin)
      tags:
        - Organizations
    delete:
      operationId: OrganizationController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Organization ID
          schema:
            type: string
      responses:
        '204':
          description: Organization deleted
        '403':
          description: Admin role required
        '404':
          description: Organization not found
        '409':
          description: Cannot delete organization with existing employees
      security:
        - bearer: []
      summary: Delete an organization (admin only)
      tags:
        - Organizations
  /organizations/slug/{slug}:
    get:
      operationId: OrganizationController_findBySlug
      parameters:
        - name: slug
          required: true
          in: path
          description: Organization slug (e.g. "peoplify")
          schema:
            type: string
      responses:
        '200':
          description: Organization details
        '404':
          description: Organization not found
      security:
        - bearer: []
      summary: Get an organization by slug (must be the caller's own org, unless admin)
      tags:
        - Organizations
  /organizations/{id}/branding:
    get:
      description: >-
        Returns the organization branding. Missing colors/logo fall back to the
        default (Peoplify) organization values.
      operationId: OrganizationController_getEffectiveBranding
      parameters:
        - name: id
          required: true
          in: path
          description: Organization ID
          schema:
            type: string
      responses:
        '200':
          description: Effective branding details
        '404':
          description: Organization not found
      security:
        - bearer: []
      summary: Get effective branding for an organization
      tags:
        - Organizations
...
info:
  title: Peoplify API
  description: HR management platform API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateOrganizationDto:
      type: object
      properties:
        name:
          type: string
          example: Peoplify
          description: Organization display name
        slug:
          type: string
          example: peoplify
          description: URL-friendly unique identifier
        primaryColor:
          type: string
          example: '#4F46E5'
          description: Primary brand color (hex)
        secondaryColor:
          type: string
          example: '#818CF8'
          description: Secondary brand color (hex)
        logo:
          type: string
          example: https://example.com/logo.png
          description: Logo URL
        isDefault:
          type: boolean
          default: false
          description: Mark as the default (root) organization
      required:
        - name
        - slug
    UpdateOrganizationDto:
      type: object
      properties:
        name:
          type: string
          example: Peoplify
          description: Organization display name
        slug:
          type: string
          example: peoplify
          description: URL-friendly unique identifier
        primaryColor:
          type: string
          example: '#4F46E5'
          description: Primary brand color (hex)
        secondaryColor:
          type: string
          example: '#818CF8'
          description: Secondary brand color (hex)
        logo:
          type: string
          example: https://example.com/logo.png
          description: Logo URL
        isDefault:
          type: boolean
          default: false
          description: Mark as the default (root) organization
...